feat: 스케쥴 추가 POST API 부분을 최상위 컴포넌트로 이동
- 최상위 컴포넌트에서만 데이터를 관리하여, 데이터 흐름을 한 곳에서 관리가 가능하며, 하위 컴포넌트에서는 UI에 집중할 수 있음.
